requirements.txt
pysrt
python-slugify
tqdm
